FIX: Support path-based project names.
Allow to correctly manage a hierarchy path project
name in the request URI. Previously only the last
part of the URI request after the forward slash
was taken into account; this was breaking in case
of path-based project names.

Example:
myprojects/project1 as project name => generated
a "cannot open repository" in branch-network.

Change-Id: Iec94112b9a807bea4661ea7a1f7a7ca88e80e18b
Signed-off-by: Luca Milanesio <luca.milanesio@gmail.com>